feat: add component RelatedImages
